fix: resolve circular depedancy on state-hooks
