+++ packages/state-hooks/src/state/wallet/hooks/index.ts
@@ -81,5 +81,4 @@ export const useTokenBalancesHook: UseTokenBalancesHookType = {
 
 export * from './hedera';
 export * from './evm';
-export * from './common';
 export * from './near';
